% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compositions.R
\name{clr_transformation}
\alias{clr_transformation}
\title{Centered Log-Ratio (CLR) transformation}
\usage{
clr_transformation(count_matrix, zero_imputation_method = "GBM")
}
\arguments{
\item{count_matrix}{Numeric matrix of counts (taxa x samples)}

\item{zero_imputation_method}{Zero count imputation criterion (see cmultRepl function): "GBM" (default), "CZM", or "pseudocount"}
}
\value{
Numeric matrix of clr-transformed abundances (taxa x samples)
}
\description{
transform compositional data in count table by (i) imputing zeros and (ii) taking the logarithm 
of each part divided by the geometric mean of all parts in the composition
}
\examples{
# Create example count data
set.seed(456)
count_data <- matrix(floor(runif(40, 0, 10)), nrow = 5, ncol = 8)
colnames(count_data) <- paste0("Sample_", 1:8)
rownames(count_data) <- paste0("Taxa_", 1:5)
 
# Convert to occurrence matrix
clr_table <- clr_transformation(count_data, zero_imputation_method="GBM")

}
