% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getmode}
\alias{getmode}
\title{Get Mode of Distribution}
\usage{
getmode(v)
}
\arguments{
\item{v}{Numeric or character vector}
}
\value{
The most frequent value in the vector
}
\description{
Calculates the mode (most frequent value) of a vector.
}
\keyword{internal}
