% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{microbiota_heatmap}
\alias{microbiota_heatmap}
\title{Microbiota Heatmap}
\usage{
microbiota_heatmap(scaled_clr_table = NULL, covariates = NULL)
}
\arguments{
\item{scaled_clr_table}{Numeric matrix of standardized clr-transformed abundance profiles (taxa x samples)}

\item{covariates}{Numeric matrix of sample(patients) covariates (samples x covariate)}
}
\value{
ggplot object
}
\description{
Plot the microbiota clr-transformed abundance profiles across samples
}
\examples{
# Example requires a completed analysis
# See package vignette for full workflow

}
