% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{reshuffle_matrix_columns}
\alias{reshuffle_matrix_columns}
\title{Reshuffle Matrix Columns}
\usage{
reshuffle_matrix_columns(Xocc)
}
\arguments{
\item{Xocc}{Binary occurrence matrix (taxa x samples)}
}
\value{
Matrix with the same dimensions but shuffled column elements
}
\description{
Randomly reshuffles the elements within each column of a matrix.
This function is used in the random null model generation.
}
\keyword{internal}
